# docs/openapi.yaml
openapi: 3.1.0
info:
  title: OmniRoute Commerce Platform API
  description: |
    Enterprise-grade B2B FMCG commerce platform combining multi-channel ordering,
    embedded finance, and gig economy logistics.
  version: 1.0.0
  contact:
    name: OmniRoute Team
    email: api@omniroute.io
  license:
    name: Proprietary
servers:
  - url: https://api.omniroute.io/v1
    description: Production
  - url: https://staging-api.omniroute.io/v1
    description: Staging
  - url: http://localhost:8080/v1
    description: Local Development
...
paths:
...
  /pricing/calculate:
    post:
      tags: [Pricing]
      summary: Calculate prices for products
      operationId: calculatePrices
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PriceRequest'
      responses:
        '200':
          description: Calculated prices
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PriceResponse'
        '400':
          $ref: '#/components/responses/BadRequest'

  # Gig Platform
...
  /payments:
    post:
      tags: [Payments]
      summary: Create a payment
      operationId: createPayment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentRequest'
      responses:
        '201':
          description: Payment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
  /wallets/{customerId}:
    get:
      tags: [Payments]
      summary: Get customer wallet
      operationId: getWallet
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Wallet details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wallet'

  # Credit
  /credit-accounts/{customerId}:
    get:
      tags: [Credit]
      summary: Get credit account
      operationId: getCreditAccount
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Credit account details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditAccount'
  /credit-scores/{customerId}:
    get:
      tags: [Credit]
      summary: Get credit score
      operationId: getCreditScore
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Credit score details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditScore'

  # SCE
...
components:
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
    PriceRequest:
      type: object
      required: [tenant_id, customer_id, items]
      properties:
        tenant_id:
          type: string
          format: uuid
        customer_id:
          type: string
          format: uuid
        items:
          type: array
          items:
            type: object
            properties:
              product_id:
                type: string
                format: uuid
              quantity:
                type: integer
    PriceResponse:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
            properties:
              product_id:
                type: string
                format: uuid
              unit_price:
                type: string
              total_price:
                type: string
              discounts:
                type: array
                items:
                  type: object
...
    PaymentRequest:
      type: object
      required: [customer_id, amount, method]
      properties:
        customer_id:
          type: string
          format: uuid
        amount:
          type: string
        method:
          type: string
          enum: [card, bank_transfer, ussd, mobile_money, wallet]
    Payment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
        amount:
          type: string
    Wallet:
      type: object
      properties:
        id:
          type: string
          format: uuid
        customer_id:
          type: string
          format: uuid
        balance:
          type: string
        status:
          type: string
    CreditAccount:
      type: object
      properties:
        id:
          type: string
          format: uuid
        credit_limit:
          type: string
        available_credit:
          type: string
        outstanding_balance:
          type: string
    CreditScore:
      type: object
      properties:
        customer_id:
          type: string
          format: uuid
        score:
          type: integer
        rating:
          type: string
          enum: [excellent, good, fair, poor]
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
security:
  - bearerAuth: []
  - apiKey: []